+++ flask_restplus/api.py
@@ -435,7 +435,7 @@ class Api(restful.Api):
                 'responses': {
                     code: (description, [fields]) if as_list else (description, fields)
                 },
-                '__mask__': True,  # Mask values can't be determined outside app context
+                '__mask__': kwargs.get('mask', True),  # Mask values can't be determined outside app context
             }
             func.__apidoc__ = merge(getattr(func, '__apidoc__', {}), doc)
             resolved = getattr(fields, 'resolved', fields)