+++ flask_restplus/swagger.py
@@ -348,15 +348,18 @@ class Swagger(object):
             params.append(param)
 
         # Handle fields mask
-        if (doc.get('__mask__') or doc[method].get('__mask__')
-                and current_app.config['RESTPLUS_MASK_SWAGGER']):
-            params.append({
+        mask = doc.get('__mask__') or doc[method].get('__mask__')
+        if (mask and current_app.config['RESTPLUS_MASK_SWAGGER']):
+            param = {
                 'name': current_app.config['RESTPLUS_MASK_HEADER'],
                 'in': 'header',
                 'type': 'string',
                 'format': 'mask',
                 'description': 'An optionnal fields mask',
-            })
+            }
+            if isinstance(mask, string_types):
+                param['default'] = mask
+            params.append(param)
 
         return params